Improve Unix process metric sampling
Take an initial CPU-time sample before asynchronous waiting can reap fast processes, document the sampled semantics, and extract the shared measurement helper.


Copilot-Session: 88ff235f-a9d7-4356-a8e7-7d4edcdcb65b
